# tema-middleware/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/routes": {
      "get": {
        "operationId": "RoutesController_list",
        "summary": "List route headers",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "site",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteListResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Routes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/routes/{xdrn}": {
      "get": {
        "operationId": "RoutesController_getByXdrn",
        "summary": "Get a route header with its detail lines",
        "parameters": [
          {
            "name": "xdrn",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteDetailResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Routes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "TEMA Middleware API",
    "description": "Enterprise integration middleware platform. Foundation endpoints only (health, readiness, version, integration health, identity). External business integrations (FSM, FSM Scheduler, Sage X3, SQL, WorkSuite, Lead Perfection) are NOT implemented yet.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Bearer token issued by the configured identity provider."
      }
    },
    "schemas": {
...
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "INTERNAL_ERROR",
            "description": "Stable, machine-readable error code."
          },
          "message": {
            "type": "string",
            "example": "An unexpected error occurred",
            "description": "Human-readable, safe error message (no internal details)."
          },
          "requestId": {
            "type": "string",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
            "description": "Correlation / request id for tracing this call."
          },
          "details": {
            "type": "array",
            "description": "Optional field-level validation details.",
            "items": {
              "type": "object"
            }
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ]
      },
...
      "RouteHeaderDto": {
        "type": "object",
        "properties": {
          "xdrn": {
            "type": "string",
            "example": "RT-USA01-0001"
          },
          "status": {
            "type": "number",
            "example": 1
          },
          "routeDate": {
            "type": "string"
          },
          "technicianId": {
            "type": "string"
          },
          "technicianName": {
            "type": "string"
          },
          "site": {
            "type": "string"
          },
          "trip": {
            "type": "number"
          },
          "createdBy": {
            "type": "string"
          }
        },
        "required": [
          "xdrn"
        ]
      },
      "RouteListResponse": {
        "type": "object",
        "properties": {
          "routes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RouteHeaderDto"
            }
          },
          "count": {
            "type": "number",
            "example": 1
          }
        },
        "required": [
          "routes",
          "count"
        ]
      },
      "RouteDetailResponse": {
        "type": "object",
        "properties": {
          "xdrn": {
            "type": "string",
            "example": "RT-USA01-0001"
          },
          "status": {
            "type": "number",
            "example": 1
          },
          "routeDate": {
            "type": "string"
          },
          "technicianId": {
            "type": "string"
          },
          "technicianName": {
            "type": "string"
          },
          "site": {
            "type": "string"
          },
          "trip": {
            "type": "number"
          },
          "createdBy": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        },
        "required": [
          "xdrn",
          "details"
        ]
      }
    }
  }
}